Fix GTK4 rendering pipeline, 64-bit portability, and crash on startup
Critical fixes ported from lrose-soloii working version:
- Add initial plot trigger in draw function (GTK4 has no configure_event)
- Use sii_really_xfer_images_cairo() for proper cairo rendering path
- Fix color table RGB conversion (GdkRGBA floats * 255 + 0.5, not >> 8)
- Reset colorize_count before queue_draw for fresh data colorization
- Add NULL safety guards throughout (pd, pal, orig_pal, wwptr, lead_sweep)
- Fix right-click support (gtk_gesture_single_set_button 0 for all buttons)
- Set parameter->changed=YES on init so color table builds on first load
- Fix 64-bit portability: long -> int32_t for rotation angle table indexing
- Fix buffer overflows: strcpy -> strncpy for 4-byte name_struct fields
- Remove paste_clipboard signal connections (not available in GTK4)
- Implement GtkTextBuffer cursor position and GdkClipboard copy
- Implement GtkFileDialog for file selection (replacing stub)
- Implement cairo PNG export for screenshots (replacing stub)
- Add widget visibility and window shrink for config layout changes
- Fix click handling with proper NULL checks and count comparison

+++ translate/swp_file_acc.c
@@ -1383,7 +1383,7 @@ int dd_create_dynamic_rotang_info(DGI_PTR dgi)
   int angle_ndx_size = 480;
   int ii, mm, size;
   int a_offset;
-  long *ang_index;
+  int32_t *ang_index;
   struct rot_ang_table *rat;
   struct rot_table_entry *entry;
 
@@ -1434,7 +1434,7 @@ int dd_create_dynamic_rotang_info(DGI_PTR dgi)
   size = sizeof(struct rot_ang_table);
   size = ((size -1)/8 +1) * 8; /* start tables on an 8 byte boundary */
   mm = size;
-  size += (angle_ndx_size * sizeof(long) +
+  size += (angle_ndx_size * sizeof(int32_t) +
            nrays * sizeof(struct rot_table_entry));
 
   /* alloc rat */
@@ -1453,7 +1453,7 @@ int dd_create_dynamic_rotang_info(DGI_PTR dgi)
   rat->sizeof_struct = size;
   rat->angle2ndx = (float)angle_ndx_size/360.0;
   rat->ndx_que_size = angle_ndx_size;
-  rat->first_key_offset = mm + angle_ndx_size * sizeof(long);
+  rat->first_key_offset = mm + angle_ndx_size * sizeof(int32_t);
   rat->angle_table_offset = mm;
   rat->num_rays = nrays;
 
@@ -1470,7 +1470,7 @@ int dd_create_dynamic_rotang_info(DGI_PTR dgi)
   /* now compute the angle indexes */
 
   a_offset = rat->angle_table_offset;
-  ang_index = (long *) ((char *) rat + a_offset);
+  ang_index = (int32_t *) ((char *) rat + a_offset);
   for (ii = 0; ii < angle_ndx_size; ii++) {
     ang_index[ii] = -1;
   }
